+++ tests/test_endpoints/test_heat_circuit_section.py
@@ -4,6 +4,7 @@ from aioresponses.core import aioresponses
 from yarl import URL
 
 from keba_keenergy_api.api import KebaKeEnergyAPI
+from keba_keenergy_api.constants import HeatCircuitExcessEnergyAvailable
 from keba_keenergy_api.constants import HeatCircuitHasMixer
 from keba_keenergy_api.constants import HeatCircuitHasPump
 from keba_keenergy_api.constants import HeatCircuitHasReturnFlowTemperature
@@ -604,6 +605,55 @@ class TestHappyPathHeatCircuitSection:
             mock_keenergy_api.assert_called_once_with(
                 url="http://mocked-host/var/readWriteVars",
                 data='[{"name": "APPL.CtrlAppl.sParam.heatCircuit[0].param.useExcessEnergy", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
+    @pytest.mark.asyncio
+    @pytest.mark.parametrize(
+        ("human_readable", "payload_value", "expected_value"),
+        [
+            (True, "true", "on"),
+            (False, HeatCircuitExcessEnergyAvailable.ON.value, 1),
+            (True, "false", "off"),
+            (False, HeatCircuitExcessEnergyAvailable.OFF.value, 0),
+        ],
+    )
+    async def test_get_get_excess_energy_available(
+        self,
+        human_readable: bool,  # noqa: FBT001
+        payload_value: str,
+        expected_value: str,
+    ) -> None:
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.heatCircuit[0].values.useExcessEnergy",
+                        "attributes": {
+                            "formatId": "fmtExcessEnergyInfo",
+                            "longText": "Excess energy HC 1",
+                            "unitId": "Enum",
+                            "upperLimit": "2",
+                            "lowerLimit": "0",
+                        },
+                        "value": payload_value,
+                    },
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: int | str = await client.heat_circuit.get_excess_energy_available(human_readable=human_readable)
+
+            assert isinstance(data, (int | str))
+            assert data == expected_value
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.heatCircuit[0].values.useExcessEnergy", "attr": "1"}]',
                 method="POST",
                 auth=None,
                 ssl=False,