+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `set_excess_energy_heating_limit_night()` to the heating circuit
 - Added endpoint `get_excess_energy_cooling_limit_night()` to the heating circuit
 - Added endpoint `set_excess_energy_cooling_limit_night()` to the heating circuit
+- Added endpoint `get_excess_energy_available()` to the heating circuit
 - Added endpoint `get_use_excess_energy()` to the heating circuit
 - Added endpoint `set_use_excess_energy()` to the heating circuit